runSearch: adapting the fixed size datatypes like uint32_t
also changed the kernel to so it can be used with those types

+++ kernels/bfs.cl
@@ -2,13 +2,13 @@
 ///
 /// This is part 1 of the Kernel from Algorithm 4 in the paper
 ///
-kernel void bfs_stage(global read_only int *vertexArray, global read_only int *edgeArray,
-                          global int *frontierArray, global int *visitedArray,
-                          global int *costArray, global int *toVisitNext,
-                          global int *previousVertexArray,
-                          int vertexCount, int edgeCount)
+kernel void bfs_stage(global read_only uint* vertexArray, global read_only uint* edgeArray,
+                          global uint* frontierArray, global uint* visitedArray,
+                          global uint* costArray, global uint* toVisitNext,
+                          global uint* previousVertexArray,
+                          ulong vertexCount, ulong edgeCount)
 {
-    int tid = get_global_id(0);
+    ulong tid = get_global_id(0);
 
     //  only process the vertex when it's in the frontier
     if (frontierArray[tid] != 0)
@@ -17,17 +17,17 @@ kernel void bfs_stage(global read_only int *vertexArray, global read_only int *e
         frontierArray[tid] = 0;
         visitedArray[tid]  = 1;
 
-        int edgeStart = vertexArray[tid];
-        int edgeEnd;
+        uint edgeStart = vertexArray[tid];
+        uint edgeEnd;
         if (tid + 1 < (vertexCount))
             edgeEnd = vertexArray[tid + 1];
         else
             edgeEnd = edgeCount;
 
-        for (int edge = edgeStart; edge < edgeEnd; ++edge)
+        for (uint edge = edgeStart; edge < edgeEnd; ++edge)
         {
             //  check if vertex behin edge was visited
-            int nextVertexId = edgeArray[edge];
+            uint nextVertexId = edgeArray[edge];
 
             if (visitedArray[nextVertexId] == 0)
             {
@@ -40,13 +40,13 @@ kernel void bfs_stage(global read_only int *vertexArray, global read_only int *e
 }
 
 
-kernel void bfs_sync(global read_only int *vertexArray, global read_only int *edgeArray,
-                          global int *frontierArray, global int *visitedArray,
-                          global int *costArray, global int *toVisitNext,
-                          global int *previousVertexArray,
-                          int vertexCount, int edgeCount)
+kernel void bfs_sync(global read_only uint* vertexArray, global read_only uint* edgeArray,
+                          global uint* frontierArray, global uint* visitedArray,
+                          global uint* costArray, global uint* toVisitNext,
+                          global uint* previousVertexArray,
+                          ulong vertexCount, ulong edgeCount)
 {
-    int tid = get_global_id(0);
+    ulong tid = get_global_id(0);
 
     if (toVisitNext[tid] != 0)
     {
@@ -58,12 +58,12 @@ kernel void bfs_sync(global read_only int *vertexArray, global read_only int *ed
 
 // Kernel to initialize buffers
 //
-kernel void bfs_init(global int *frontierArray, global int *visitedArray,
-                     global int *costArray,
-                     global int *previousVertexArray,
-                     int vertexCount, unsigned int sourceVertex)
+kernel void bfs_init(global uint* frontierArray, global uint* visitedArray,
+                     global uint* costArray,
+                     global uint* previousVertexArray,
+                     ulong vertexCount, uint sourceVertex)
 {
-    unsigned int tid = get_global_id(0);
+    uint tid = get_global_id(0);
 
     if (sourceVertex == tid)
     {
